fix: refactored currentGame so playerPanels are rendered separately from the Phaser game, allowing Phaser to send a signal to render the challengeModal
