+++ public/input.scss
@@ -109,23 +109,38 @@ input {
 
 #gameView {
   display: flex;
-  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+#panelsBox {
+  display: flex;
+  justify-content: space-between;
+  width: 95vw;
 }
 .leftside {
   display: flex;
   flex-direction: column;
   margin-right: 2rem;
+  margin-left: 2rem;
 }
-#theGame {
-  display: flex;
+canvas {
   border-style: solid;
   border-color: black;
   border-width: 5px;
+  position: absolute;
+  top: 11rem;
+  left: 23%;
+}
+#insideTheGame {
+  position: absolute;
+  top: 100px;
+  left: 100px;
 }
 .rightside {
   display: flex;
   flex-direction: column;
   margin-left: 2rem;
+  margin-right: 2rem;
 }
 .singlePlayerBox {
   display: flex;
@@ -213,6 +228,16 @@ input {
   display: flex;
   flex-direction: column;
   align-content: center;
+  position: absolute;
+  top: 40%;
+  left: 33%;
+  z-index: 1;
+  padding: 2rem;
+  background-color: #a39bb0;
+  border-style: ridge;
+  border-width: 1rem;
+  border-color: #fef2eb;
+  box-shadow: 1px 1px 400px 80px black;
 }
 .answerChoices {
   display: flex;
